requirements_base.txt
otree-core==0.5.2
Django==1.8.8
